refactor: extract cli option parsing
