+++ crates/mcp-compressor-core/src/app/mod.rs
@@ -1,1 +1,2 @@
+pub mod options;
 pub mod paths;
